+++ momoim-external-api/src/main/java/com/triplem/momoim/api/gathering/controller/GatheringController.java
@@ -5,8 +5,8 @@ import com.triplem.momoim.api.common.DefaultApiResponse;
 import com.triplem.momoim.api.gathering.dto.GatheringCategoryInformation;
 import com.triplem.momoim.api.gathering.dto.GatheringDetail;
 import com.triplem.momoim.api.gathering.dto.GatheringListItem;
-import com.triplem.momoim.api.gathering.dto.ModifyGatheringRequest;
-import com.triplem.momoim.api.gathering.dto.RegisterGatheringRequest;
+import com.triplem.momoim.api.gathering.request.ModifyGatheringRequest;
+import com.triplem.momoim.api.gathering.request.RegisterGatheringRequest;
 import com.triplem.momoim.api.gathering.service.GatheringJoinService;
 import com.triplem.momoim.api.gathering.service.GatheringManagementService;
 import com.triplem.momoim.api.gathering.service.GatheringService;
